requirements.txt
django-threadlocals
pymysql
click
requests
xlrd
python-alipay-sdk